Native bridge bug fixes + directory restructure
