+++ ForemWebView/ForemWebView+WKScriptMessageHandler.swift
@@ -1,5 +1,9 @@
 import WebKit
 
+enum BridgeMessageType {
+    case podcast, video
+}
+
 extension ForemWebView: WKScriptMessageHandler {
     public func userContentController(_ userContentController: WKUserContentController,
                                       didReceive message: WKScriptMessage) {
@@ -14,6 +18,31 @@ extension ForemWebView: WKScriptMessageHandler {
             guard let hapticType = message.body as? String else { return }
             handleHapticMessage(type: hapticType)
         default: ()
+        }
+    }
+
+    // Helper function that will send Bridge messages into the DOM
+    internal func sendBridgeMessage(type: BridgeMessageType, message: [String: String]) {
+        var jsonString = ""
+        let encoder = JSONEncoder()
+        if let jsonData = try? encoder.encode(message) {
+            jsonString = String(data: jsonData, encoding: .utf8) ?? ""
+        }
+
+        var javascript = ""
+        // Supported messages
+        switch type {
+        case .podcast:
+            javascript = "document.getElementById('audiocontent').setAttribute('data-podcast', '\(jsonString)')"
+        case .video:
+            javascript = "document.getElementById('video-player-source').setAttribute('data-message', '\(jsonString)')"
+        }
+
+        guard !javascript.isEmpty else { return }
+        evaluateJavaScript(wrappedJS(javascript)) { _, error in
+            if let error = error {
+                print("Error sending Podcast message (\(message)): \(error.localizedDescription)")
+            }
         }
     }
