+++ ForemWebView/JavaScript & CSS/bodyMutationObserver.js
