+++ ForemWebView/JavaScript & CSS/fetchForemInstanceMetadata.js
