+++ ForemWebView/JavaScript & CSS/invertedImages.css
