+++ ForemWebView/Media Manager/ForemMediaManager+LockedScreenControls.swift
